metiscoin now has valid shares again, finally cleaned up how to do data alignment

+++ xptMiner/shavite.c
@@ -1817,7 +1817,7 @@ sph_shavite512_addbits_and_close(void *cc, unsigned ub, unsigned n, void *dst)
 	^ AES2[((t1) >> 16) & 0xFF] \
 	^ AES3[((t2) >> 24)];
 
-void shavite_big_core_opt(long long * msg, long long * dataOut)
+void shavite_big_core_opt(unsigned int* msg, unsigned int* dataOut)
 {
 	sph_u32 kt;
 	sph_u32 p0, p1, p2, p3, p4, p5, p6, p7;
